requirements.txt
alembic==1.14.1
annotated-types==0.7.0
anyio==4.8.0
app==0.0.1
appdirs==1.4.4
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
asgiref==3.8.1
attrs==24.2.0
blessed==1.20.0
blinker==1.9.0
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
distlib==0.3.9
Django==5.1.4
fastapi==0.115.6
filelock==3.17.0
Flask==3.1.0
Flask-Migrate==4.1.0
Flask-SQLAlchemy==3.1.1
flyingcircus==0.1.4.1
greenlet==3.1.1
h11==0.14.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.5
Mako==1.3.8
MarkupSafe==3.0.2
numpy==2.2.2
outcome==1.3.0.post0
packaging==24.2
pandas==2.2.3
pillow==11.1.0
platformdirs==4.3.6
playwright==1.49.1
pycparser==2.22
pydantic==2.10.5
pydantic_core==2.27.2
pyee==12.0.0
PySocks==1.7.1
python-dateutil==2.9.0.post0
pytk==0.0.2.1
pytz==2024.2
requests==2.32.3
selenium==4.27.1
setuptools==75.6.0
setuptools-scm==8.1.0
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
SQLAlchemy==2.0.37
sqlparse==0.5.3
starlette==0.41.3
trio==0.27.0
trio-websocket==0.11.1
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.2.3
virtualenv==20.29.1
wcwidth==0.2.13
websocket-client==1.8.0
Werkzeug==3.1.3
wsproto==1.2.0
